move all context and modify corresponding import path
