--- src/context/DataContext.js
@@ -1,4 +1,0 @@
-import { createContext, useState } from "react";
-const DataContext = createContext(null);
-
-export default DataContext
