--- src/context/PopupContext.js
@@ -1,4 +1,0 @@
-import { createContext } from "react";
-const PopupContext = createContext(null);
-
-export default PopupContext
