+++ src/module/core/context/WindowResize.js
