+++ src/page/Insight.jsx
@@ -7,7 +7,7 @@ import { Line } from 'react-chartjs-2';
 import { Chart, registerables } from 'chart.js';
 
 import { useContext } from 'react';
-import DataContext from '../context/DataContext';
+import DataContext from '../module/core/context/DataContext';
 import * as XLSX from '@e965/xlsx';
 import { saveAs } from 'file-saver';
 import ExcelImage from '../module/core/asset/image/excel.png'
