+++ src/page/gameTheory/Input.jsx
@@ -7,14 +7,14 @@ import ExcelImage from '../../module/core/asset/image/excel.png'
 import { saveAs } from 'file-saver';
 import * as XLSX from '@e965/xlsx';
 import { useContext } from 'react';
-import DataContext from "../../context/DataContext"
+import DataContext from "../../module/core/context/DataContext"
 
 import { useNavigate } from 'react-router-dom';
 import { Link } from 'react-router-dom'
 
 import Loading from '../../module/core/component/Loading';
 import MaxMinCheckbox from '../../module/core/component/MaxMinCheckbox'
-import PopupContext from '../../context/PopupContext';
+import PopupContext from '../../module/core/context/PopupContext';
 export default function InputPage() {
     //initialize form data
     const [excelFile, setExcelFile] = useState(null);
