+++ src/page/stableMatching/Output.jsx
@@ -2,14 +2,14 @@ import React from "react";
 import "../../module/stableMatching/css/output.scss";
 import GraphImage from "../../module/core/asset/image/graph.png";
 import { useContext, useState } from "react";
-import DataContext from "../../context/DataContext";
+import DataContext from "../../module/core/context/DataContext";
 import { useNavigate } from "react-router-dom";
 import NothingToShow from "../../module/core/component/NothingToShow";
 import Loading from "../../module/core/component/Loading";
 import Popup from "../../module/core/component/Popup";
 import axios from "axios";
 import ParamSettingBox from "../../module/core/component/ParamSettingBox";
-import PopupContext from "../../context/PopupContext";
+import PopupContext from "../../module/core/context/PopupContext";
 import * as XLSX from '@e965/xlsx';
 import SockJS from "sockjs-client";
 import { v4 } from "uuid";
