+++ src/page/stableMatching/Processing.jsx
@@ -3,11 +3,11 @@ import "../../module/stableMatching/css/processing.scss";
 import {useNavigate} from "react-router-dom";
 import {useContext, useState, useEffect} from "react";
 import axios from "axios";
-import DataContext from "../../context/DataContext";
+import DataContext from "../../module/core/context/DataContext";
 import NothingToShow from "../../module/core/component/NothingToShow";
 import Loading from "../../module/core/component/Loading";
 import ParamSettingBox from "../../module/core/component/ParamSettingBox";
-import PopupContext from "../../context/PopupContext";
+import PopupContext from "../../module/core/context/PopupContext";
 
 //TODO: algorithm selection
 export default function InputProcessingPage() {
